src/main/java/com/epnj/intelligentpoint/api/utils/PasswordUtils.java
package com.epnj.intelligentpoint.api.utils;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;

public class PasswordUtils {
    private static final Logger log = LoggerFactory.getLogger(PasswordUtils.class);

    public PasswordUtils() {
    }

    public static String generateCrypt(String password) {
        if (password == null) {
            return password;
        }
        log.info("Generating hash with the BCrypt");
        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
        return bCryptPasswordEncoder.encode(password);
    }

}
